Make WDL submission id match with other files submission id
Closes #221

The WdlFile class had submission_id as a kwarg. It was not properly being used in submission method of the CLI, which caused it to generate its own submission id.
This fixes the issue so that the submission_id generated when the files are moved and compressed match. This also improves the error message when validating the inputs.json file. It also includes a renaming of some variables for better readability and understanding.

Flake8 fixes

Skip test if no WOMTool present

This test unfortunately requires WOMTool to be present

Add helper tools (eg WOMTool) to PATH

Add debug print statements

Require WOMTool to be in PATH

Remove debug statements

Remove debug statements