Fix wfcopy for subworkflows that don't use sharding
Detect whether a cromwell directory contains a subworkflow but is not a shard dir. If found, copy to output dir. Closes #196

Related (also bug in subworkflow traversal), and being addressed in showing errors files for subworkflow tasks. Closes #195

This MR includes additional tests for testing the client functionality. This also refactors workflow.py to simplify the code. The idea here is that there are now two classes WdlFile and WorkflowInputs that represent the wdl and inputs.json that are submitted by the user. Any operations to query information about a workflow are separated amongst the two rather than a single Workflow class.

The bigger change is eliminating the Workflow class and removing the need to require `womtool inputs` to be ran.

Eventually this should be brought back to perhaps satisfy #147.

I think in this effort I can maybe start moving functions into a client library since that would help with testing.