Raise WdlError if stderr present
Fixes #284

Validation would only check if the stderr from WOMTool would just show a message of subworkflows missing.
This would mean that if we submit with a Wdl file with bad syntax, the submission would go on without
notifying the user that their Wdl file is incorrect.
The validation function should check if the stderr message has a missing subworkflows message and if not
raise the WdlError with the original message.