Modify container exec script arguments
Closes #546

The arguments to JTM are passed incorrectly when executing the shifter_exec.sh script.
This causes Cromwell to throw error code 79. Both singularity and shifter container run
times require the refdata database to be mounted into the container. Rather than pass
this is as an argument to JTM, have this argument present in the exec scripts.

Add same number of arguments to both wrappers

This moves $SITE_REF_DATA_DIR and /refdata as parameters that will be
passed in. This will make it so that both singularity and shifter will
expect a total number of 5 arguments.