Add recursive traversal of JSON
Closes #584

Adds back some of the old functions that were deleted by commit: 08456cb38a064b95313ae260ef9e066d1aa425ba.
In the WorkflowInputs constructor, the JSON is recursively traversed to check whether filepaths are relative.
If they are, they are converted to absolute filepaths. It also includes some minor refactors like reordering
the location of methods so that it follows the order of: public, property, private methods to follow best
practices.

Also includes a bugfix in the `apply_filepath_op` function. It would return None rather than
the actual element itself. This is because if the inner `if` conditional fails, it would return
None rather than exit with the outer else statement.