Use env vars to set location of user directories
Issue #636

Modifies the CI/CD deployment scripts so that the user directories of
staging_dir and data_repo are created with the sticky bit set for groups. This
also makes sure to add to the configuration directory the required
settings to properly point to these locations. Additionally, this MR
includes a utils.sh script which has all the common functions present
there. The deployment scripts should eventually move there.