add globus confidential app to config and switch transfers from native to conf app
edit for flake8

update central's tests

update site tests

rename function

make function to authorize globus transfer client

removed unused function and use config.conf directly

rename function

add user info query back

Remove NA and CA from client id's and secret

Since we are focusing on the client application way of authenticating we should
remove all instances of NA_CLIENT_ID and CA_CLIENT_ID to avoid confusion.
For now we are only focusing on a single client id.

Change path of import file

The harcoded path for the JAWS configuration file made it difficult
to run tests unless located above the "tests" directory. This made
it harder to run in debug mode via IDE's like PyCharm. This changes
the path so that it can be the correct path in any location.

Eventually this needs to change as putting imports in __init__.py
is unnecessary especially under the tests directory. We do not require
the __init__.py in tests as it should not be a module that needs to be imported.
We also need to refactor the Config class so that globals are not used.

Change root_dir and default_dir

This changes the root_dir and default_dir to point to root.
The `test_daemon.py:test_transfer_results` test was
erroneously giving a passing test because the transfer would error out
on the checking of the globus basepath and would return None. The class
mocks were hiding the issue.

Remove underscore and add docstring

_method names are mainly use to denote private functions in classes. Since
this is not a class it is not necessary to have that. Also adds a docstring to
the function.

Modify configuration and remove transfer client

MockTransferClient is not longer needed since we have a GlobusService
that can be mocked out. This also changes the variable names for
client_id and client_secret.

Add GlobusService class

This class takes care of the connection and transfer logic that is used
with the globus_sdk. Clients will interact with this class to perform
any related globus tasks.

Modify daemon to use GlobusService

This modifies the Daemon to have a GlobusService attribute that
can be used to perform globus related tasks. It removes some of the
globus logic from daemon and moves it to globus.py. Tests are also
added.

Add updated EMSL endpoint

Modify CLIENT_ID names

Removes the NA/CA prefix. Sticks to one naming convention for now.