Modify paths based off of shared endpoint path
Closes #721

Modifies the deployment scripts to use `SITE_GLOBUS_HOST_PATH`. This will correspond to the
root host path of the shared endpoint. The shared endpoint can be either root (/) or a
path on some filesystem (eg /global/scratch/jaws). Gets rid of the ROOT_DIR and DEFAULT_DIR
since we are using shared endpoints that will contain their own root path.

This also includes changes throughout site, client, and central code
that reflect these modifications.

Use globus_host_path in configuration classes and clients

Updates the config code to use the new `globus_host_path` configuration option. This also updates
the code to account for the configuration changes so that they pass.

- `test_analysis.py`: The configuration fixture was moved to conftest.py since it was using
two different configuration fixtures. Now they are using the same generated configuration file.
This also slightly modifies the list_sites test to test whether the right set of sites are included
in the config file.
- `conftest.py`: The configuration file fixture is modified to match our most current schema.
- `test_config.py`: changes some of the configuration options so that they correctly match each other
in the tests.
-`jaws-central.ini`: modified to look like the most current schema.

Change config params and fix tests

This changes the code in client to adjust for the new configuration
name change. It also fixes failing tests that were happening locally
on a Macbook with python version 3.7.4.

- `analysis.py`: Removed the out_endpoint option for the user since it is no longer needed.
This should likely be abstracted away by the user since they do not need to rely on remembering
globus endpoints. The setup should be done my administrators. Also removed code that relied on
user input of an output directory. The user no longer needs to specify this output since it is
controlled by JAWS. Changes some variable names due to the configuration changes.

- `workflow.py`: Tests were failing locally on a Macbook with a JAWS virtualenv with python 3.7.4.
The result was that the original input to subprocess.run was included with stdout. This caused the
subworkflows tests to fail since it could not find the right file. This adds a regex check to make
sure that the output does not resemble a string input to subprocess.run. Comments included as well.

- `conftest.py` and `test_analysis.py`: Changed the configuration in conftest to look like what is
out in production. Modified some tests so that they account for these configuration updates.

Create GlobusService class to manage transfers

This moves the logic of submitting transfers out of analysis.py and into a GlobusService
class that can submit and modify the paths according to their endpoints. We also collect
the host paths from the configuration file and modify and create virtual paths.

This includes tests for the new virtual_transfer_path method.

Add virtual path method to site's Globus class

Includes updates to configuration changes. The virtual path method now
makes sure that the paths from the compute site have the correct virtual path
for Globus. The destination dir is left alone and is already modified.

Modify output dir to Globus virtual paths

In the case where the endpoint host_path is non root, the path should
be modified so that we properly return the files back. This is done at central
since the source and destination host paths are already known. We then insert this
back into the database. Site then only needs to query the output directory without
needing to modify the path and know the return site's host path.

Also includes additional paths that match this scenario in a test.