Setfacl on site directories
Closes #800

Sets the appropriate ACL rules for directories at a site. This is done
through shims since parallel filesystems fails to acknowledge filesystem
operations when done on a different filesystem. This is now done with
deployments to site rather than with client.

Add site var DATA_REPO_BASEDIR

escape $ for variable

Add chmod 2775 in script

Removed unused gitlab variable

We no longer need the staging dir since we now stage to where the inputs
directory is located. The path is determined in the `define-env` script