SetACLs on existing directories
Closes #931

Modifies the cromwell script so that permissions are correctly set to
directories that already exist in the cromwell-executions directory.
The current script only sets ACLs for parent directories so that any
newly created directories will have the correct ACLs. We found however
that the script does not fix the permissions for direcetories that
already exist.

Add shopt to avoid loops on empty directory

Remove unnecessary loop

The recursive flag should take care of traversing down the directory
tree