Create console scripts for recv.py and send.py
Closes #1077

Modifies setup.py for parsl so that jaws-parsl-send and jaws-parsl-recv
are console scripts that can be used for running parsl workflows. This
also changes the shim so that it invokes the correct script.

tRefactor parsl config imports

Refactors how the configs are imported. Each site config is now its own
module and separate file. Adds a new local backend for laptop testing.

Add missing parsl_configs