Parallelize file transfers with parallel_sync
Closes #1642

Uses the python package parallel_sync to paralallize file transfers.
The number of threads are set dynamically depending on the amount of
files in the file transfers. This max threads can be set in the config
file.

In order to allow for parallel_sync to be installed, the python version
was downgraded to 3.10. A dependency of parallel_sync does not currently
support 3.11.