feat(pubsub): add background tasks
Background tasks warp functions into "tasks" to be
performed outside of the event loop and in separate thread, preventing
the task from blocking IO from execution. This is especially useful
for executing IO functions that do not use an asynchronous library or
long compute heavy functions; the former use case is the most likely
use case for this.