chore(build): add image_version.yaml inside images
Creates an image_version.yaml that is located inside the
container. This can be used to determine whether the correct
code is deployed in the container.

A dedicated step used to generate the image_version.yml is added
to avoid having a dependency inside the container for git.

Additionally, a default image_version.yml file is added to the repo.
This can be updated automatically by using the build.sh.

Closes: #1733