feat(transfers): Add new transfer function
`parallel-sync` was rewritten in version 2.0; this new version did not
include a local transfer function. Futhermore pinning parallel-sync to
its last version before the rewrite made it incompatible with
python3.11, as it requires a package called `brunch` that has not been
updated since 2011 and uses deprecated features. For these reasons we
have rewritten the transfer local copy to eliminate our dependency on
parallel-sync.