fix(transfers): add retry logic
Parallel local file copy was recently changed to use python's built in
`shutil.copy` function instead of a library that makes an external
call to rsync. Since the release of python 3.8 the built in copy
utility has been rewritten to use kernel level fast copy calls so it
should be as fast if not faster than rsync to copy files locally.